Introduce a new PageCache object
and split out get_post_fields into its own utility function.
